+++ client/src/components/Chat/Chat.jsx
@@ -98,4 +98,3 @@ const Chat = () => {
 }
 
 export default Chat;
-
