+++ client/src/components/SearchUser/OnClickOnStart/OnClickOnStart.jsx
@@ -1,12 +1,19 @@
-import React from 'react'
-import { useState, useContext } from 'react'
+import React, { useState, useContext, useEffect } from 'react'
 import { useSelector } from 'react-redux';
+import { SocketContext } from '../../../Context';
 
 
 import FoundUser from '../FoundUser/FoundUser';
 
 const OnClickOnStart = () => {
   const [isFound, setIsFound] = useState(false);
+  const socket = useContext(SocketContext)
+
+  const userID = useSelector(state => state.ID)
+
+  useEffect(() => {
+    socket.connected && socket.emit('adding', ({ userID }))
+  }, [socket, userID])
 
   return (
     isFound ? <FoundUser /> : <div>Searching.....</div>
